+++ CHANGELOG.md
@@ -25,6 +25,8 @@ speaking, considered to be the player's experience.
   account other defenders present, showing too high when other units were
   present.
 - Fixed the estimated assassination success chance against a king being too low.
+- Fixed the estimated assassination success chance being too low based on the
+  spy skill.
 - Fixed the production queue sizes showing the wrong number when training a unit
   finishes.
 - Fixed the production queue size display bouncing down and up in multiplayer
@@ -56,6 +58,8 @@ speaking, considered to be the player's experience.
 - Adjusted the graphics of Standard tileset terrain that looked like it should
   allow building and movement around its corners and edges but didn't, so that
   it's now obvious by just looking that it's not.
+- Reworked the assassination chance estimate calculation to better take into
+  account the possibility or lack of enemy spies.
 - Disabled the caravan clone button (and remove it in the trade report) for idle
   caravans, since trying to clone them does nothing anyway.
 - Changed default keybind for train soldier to `F` from `W`, to make it more